+++ src/Notify/CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-- Added body, icon, tag ,renotify to controller.js for mercure-notifier
+-   Added body, icon, tag ,renotify to controller.js for mercure-notifier
 
 ## 2.13.2